TestPodaciPrimjeri/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pyodbc>=5.1.0
aiohttp>=3.11.0
apscheduler>=3.10.4
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
jinja2>=3.1.4
python-multipart>=0.0.12
python-dotenv>=1.0.1
python-dateutil>=2.9.0
